Refactoring oracle extractor to just one class.
